Backend: 1. Added Spring Security To the Project 2. secured open endpoints 3. configured jwt token based authetication for api's 4. created error pages 5. created api controllers for ( authentication, semesters, subjects, students, course ) 6. removed unnaccery code
Flutter App : Faculty
1. configured authetication through api's
2. implemented token based request passing machanism
3. recreated necessary models
4. reconfigured auto login

Flutter App : Student
1. only small change in constants.dart

+++ eAttendance Backend Web/src/main/java/com/attendance/system/controller/AttendanceController.java
@@ -46,7 +46,7 @@ public class AttendanceController {
 	}
 	
 	@GetMapping("fill/{mid}/{sid}")
-	public ResponseEntity<Boolean> fillAttendance(@PathVariable Integer mid,@PathVariable Integer sid){
+	public ResponseEntity<Boolean> fillAttendance(@PathVariable Integer mid,@PathVariable Long sid){
 		return attendanceService.fillAttendance(mid,sid);
 	}
 }
